Rewrote the Variant Loader to make all the development sided events, redundant
