+++ common/src/main/java/software/bluelib/BlueLibCommon.java
@@ -12,6 +12,7 @@ import software.bluelib.api.event.mod.ModIntegration;
 import software.bluelib.api.net.NetworkRegistry;
 import software.bluelib.api.utils.logging.BaseLogLevel;
 import software.bluelib.api.utils.logging.BaseLogger;
+import software.bluelib.example.event.VariantProvider;
 
 public class BlueLibCommon {
