+++ common/src/main/java/software/bluelib/api/entity/variant/IVariantProvider.java
@@ -1,0 +1,11 @@
+package software.bluelib.api.entity.variant;
+
+import java.util.List;
+
+public interface IVariantProvider {
+	List<String> getEntityNames();
+	
+	default String getBasePath() {
+		return "variant/entity/";
+	}
+}
