+++ neoforge/src/main/java/software/bluelib/event/ReloadHandler.java
@@ -1,8 +1,5 @@
-// Copyright (c) BlueLib. Licensed under the MIT License.
+package software.bluelib.event;
 
-package software.bluelib.example.event;
-
-import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 
@@ -13,35 +10,45 @@ import net.neoforged.neoforge.event.OnDatapackSyncEvent;
 import net.neoforged.neoforge.event.server.ServerStartingEvent;
 import software.bluelib.BlueLibCommon;
 import software.bluelib.BlueLibConstants;
-import software.bluelib.api.reload.ReloadEventHandler;
+import software.bluelib.api.entity.variant.IVariantProvider;
 import software.bluelib.api.utils.logging.BaseLogLevel;
 import software.bluelib.api.utils.logging.BaseLogger;
+import software.bluelib.entity.variant.VariantLoader;
 
 @EventBusSubscriber(modid = BlueLibConstants.MOD_ID)
-public class ReloadHandler extends ReloadEventHandler {
+public class ReloadHandler {
+
+    private static IVariantProvider provider;
+
+    public static void setProvider(IVariantProvider pVariantProvider) {
+        provider = pVariantProvider;
+    }
 
     @SubscribeEvent
     public static void onServerStart(ServerStartingEvent pEvent) {
+        if (provider == null) return;
+
         BlueLibConstants.SCHEDULER = new ScheduledThreadPoolExecutor(1);
         BlueLibConstants.server = pEvent.getServer();
-        ReloadHandler.LoadEntityVariants(pEvent.getServer());
+        loadEntityVariants(pEvent.getServer());
         BaseLogger.log(BaseLogLevel.INFO, BlueLibCommon.Translation.log("variants.loaded"), true);
     }
 
     @SubscribeEvent
     public static void onDatapackSync(OnDatapackSyncEvent pEvent) {
-        ReloadHandler.LoadEntityVariants(pEvent.getPlayerList().getServer());
+        if (provider == null) return;
+
+        loadEntityVariants(pEvent.getPlayerList().getServer());
         BaseLogger.log(BaseLogLevel.INFO, BlueLibCommon.Translation.log("variants.reloaded"), true);
     }
 
-    private static final String basePath = "variant/entity/";
+    private static void loadEntityVariants(MinecraftServer pServer) {
+        List<String> entityNames = provider.getEntityNames();
+        String basePath = provider.getBasePath();
 
-    private static final List<String> ENTITY_NAMES = Arrays.asList("exampleone", "exampletwo");
-
-    public static void LoadEntityVariants(MinecraftServer pServer) {
-        for (String entityName : ENTITY_NAMES) {
+        for (String entityName : entityNames) {
             String folderPath = basePath + entityName;
-            ReloadEventHandler.registerEntityVariants(folderPath, pServer, BlueLibConstants.MOD_ID, entityName);
+            VariantLoader.loadVariants(folderPath, pServer, entityName);
             BaseLogger.log(BaseLogLevel.INFO, BlueLibCommon.Translation.log("variants.loaded.entity", entityName), true);
         }
     }
